+++ packages/core/src/components/accordion/components/accordion-section.tsx
@@ -26,7 +26,7 @@ import {
 import { ExpandableAnimationEndEvent } from "../../expandable/expandable";
 
 // eslint-disable-next-line no-console
-const log = (window as any)["_dsoLog"] === true ? console.log.bind(console.log) : function () {};
+const log = (window as never)["_dsoLog"] === true ? console.log.bind(console.log) : function () {};
 
 const HandleElement: FunctionalComponent<{
   handleUrl: string | undefined;
@@ -299,10 +299,10 @@ export class AccordionSection implements ComponentInterface {
     });
   };
 
-  private handleExpandableAnimationStart = (e: CustomEvent<any>) => {
+  private handleExpandableAnimationStart = () => {
     this.dsoAnimationStart.emit({
       animation: this.open ? "opening" : "closing",
-      scrollIntoView: (behavior: ScrollBehavior = "auto") => this.scrollIntoView(e.detail.bodyHeight, behavior),
+      scrollIntoView: (behavior: ScrollBehavior = "auto") => this.scrollIntoView(undefined, behavior),
     });
   };